Hack to support bugs in dune-grid 2.2.0
This particular version of dune-grid tested the define ENABLE_ALUGRID
instead of HAVE_ALUGRID, which was the intention. Our build system
doesn't set this define, so we do it here before including in case we
have that version.

It should otherwise be harmless, since HAVE_ALUGRID should only be
defined in DUNE if ENABLE_ALUGRID is true (1) anyway.